# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movieflix API",
        "description": "API responsável por gerenciar os filmes do Movieflix",  
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
        "get": {
            "tags": ["Movies"],
            "summary": "Obter todos os filmes",
            "responses": {
            "200": {
                "description": "Retorna uma lista de filmes",
                "content": {
                "application/json": {
                    "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/components/schemas/Movie"
                    }
                }
                }
            }
            }
        }
        }
    },
    "/movies/{id}": {
        "put": {
            "tags": ["Movies"],
            "sumary": "Atualizar um filme",
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID do filme a ser atualizado",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MovieInput"
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Filme atualizado com sucesso"
                },
                "404": {
                    "description": "Filme não encontrado"
                },
                "500": {
                    "description": "Erro interno do servidor"
                }
            }
        },
        "delete": {
            "tags": ["Movies"],
            "sumary": "Deletar um filme",
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID do filme a ser deletado",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "Filme deletado com sucesso"
                },
                "404": {
                    "description": "Filme não encontrado"
                },
                "500": {
                    "description": "Não foi possível deletar o filme"
                }
            }
        }
    }
},
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero"
                            }
                        }
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma do filme"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do idioma"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do idioma"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Número de Oscars ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": ["title", "genre_id", "language_id", "oscar_count", "release_date"],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Número de Oscars ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
            }
        }
    }
}
}